requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
python-dotenv==0.19.0
openai==1.3.0
markdown2==2.4.0
SQLAlchemy==1.4.47
